feat(gui): enforce FutureOS sign-in with full-screen login gate
When no FutureOS account is signed in (on launch, or after signing out), a full-screen gate overlays the entire app with a title, subtitle, sign-in button, and language switch — the rest of the app is unreachable until sign-in succeeds. The existing device-code flow opens the browser directly (no pairing-code display, no modal).

Extract the login state machine from FutureLoginDialog into useFutureLoginFlow (adds idle phase, cancel, callers control begin timing). Add useFutureSignedIn (providers search + future-auth-changed event-driven silent reload) and the ForceLoginGate component. AppShell early-returns the gate (or a short neutral spinner on first load) after all hooks, so the rules-of-hooks invariant is preserved. All existing login surfaces (settings Account / Providers dialogs) still work unchanged.
