+++ gui/src/components/layout/AppShell.tsx
@@ -23,10 +23,12 @@ import { ToastHost } from "../ui/ToastHost";
 import { ActivityRail } from "./ActivityRail";
 import { AppShellDialogs } from "./AppShellDialogs";
 import { ContextPanel } from "./ContextPanel";
+import { ForceLoginGate } from "./ForceLoginGate";
 import { useAgentConnection } from "./hooks/useAgentConnection";
 import { useApprovals } from "./hooks/useApprovals";
 import { useAppSettings } from "./hooks/useAppSettings";
 import { useAutoUpgradeSkills } from "./hooks/useAutoUpgradeSkills";
+import { useFutureSignedIn } from "./hooks/useFutureSignedIn";
 import { useModelSelection } from "./hooks/useModelSelection";
 import { useNewConversation } from "./hooks/useNewConversation";
 import { useRightPanelWidth } from "./hooks/useRightPanelWidth";
@@ -69,6 +71,9 @@ export function AppShell() {
   const { appSettings, changeSettings } = useAppSettings();
   useAutoUpgradeSkills(appSettings.autoUpgradeSkills);
   const { hasUpdate, cachedStatus, markSeen: markUpdateSeen } = useUpdateChecker();
+  // Drives the forced-login gate below. Kept with the other top-level hooks so
+  // the early returns further down stay after every hook call (rules of hooks).
+  const { signedIn, initialLoading } = useFutureSignedIn();
 
   const centerRef = useRef<HTMLElement>(null);
   const {
@@ -358,6 +363,20 @@ export function AppShell() {
     onTogglePinThread: handleTogglePinThread,
     onToggleExpanded: handleToggleLeftPanel,
   };
+
+  // Forced login: until the FutureOS account is confirmed, block the whole app
+  // with the login gate. `initialLoading` covers only the first probe so a
+  // signed-in user never sees the gate flash; reloads on auth change are silent
+  // (data stays non-null), so sign-in / sign-out never flash this neutral frame.
+  if (initialLoading) {
+    return (
+      <div className="flex h-full items-center justify-center bg-canvas">
+        <span className="size-6 animate-spin rounded-full border-2 border-accent-soft border-t-accent" />
+      </div>
+    );
+  }
+  if (!signedIn)
+    return <ForceLoginGate />;
 
   return (
     <div className="relative flex h-full min-h-0 overflow-hidden bg-canvas text-ink">
