+++ gui/src/integrations/agent/providers.ts
@@ -1,3 +1,4 @@
+import { emitFutureEvent } from "../../lib/futureEvents";
 import { invokeCommand } from "../tauri/invoke";
 
 /** The FutureOS builtin provider id — its key backs the signed-in account. */
@@ -72,8 +73,10 @@ export async function updateBuiltinProviderKey(input: {
 }) {
   const view = await invokeCommand<ProvidersView>("update_builtin_provider_key", { input });
   // Setting the FutureOS key by hand (Providers page) changes the account too.
-  if (input.id === FUTURE_PROVIDER_ID)
+  if (input.id === FUTURE_PROVIDER_ID) {
     clearFutureProfileCache();
+    emitFutureEvent("future-auth-changed", undefined);
+  }
   return view;
 }
 
@@ -120,15 +123,19 @@ export async function startFutureLogin() {
 export async function pollFutureLogin(deviceCode: string) {
   const result = await invokeCommand<FutureLoginPoll>("poll_future_login", { deviceCode });
   // A completed device login writes a new key — invalidate here so it doesn't
-  // matter which page (Account or Providers) ran the login dialog.
-  if (result.status === "authorized")
+  // matter which page (Account or Providers) ran the login dialog, and tell the
+  // app-wide login gate to clear.
+  if (result.status === "authorized") {
     clearFutureProfileCache();
+    emitFutureEvent("future-auth-changed", undefined);
+  }
   return result;
 }
 
 export async function logoutFutureProvider() {
   const view = await invokeCommand<ProvidersView>("logout_future_provider");
   clearFutureProfileCache();
+  emitFutureEvent("future-auth-changed", undefined);
   return view;
 }
